Decouple NavigationManager from the way NavControllers are retrieved (from tab container layout)
