+++ app/src/main/java/com/taurus/modernandroiddevelopmentkata/di/module/ActivityBuilderModule.kt
@@ -3,7 +3,6 @@ package com.taurus.modernandroiddevelopmentkata.di.module
 import com.taurus.modernandroiddevelopmentkata.MainActivity
 import com.taurus.modernandroiddevelopmentkata.core.di.scope.ActivityScope
 import com.taurus.modernandroiddevelopmentkata.navigation.executor.NavigationExecutorModule
-import com.taurus.modernandroiddevelopmentkata.similarmovies.SimilarMoviesModule
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -13,6 +12,7 @@ abstract class ActivityBuilderModule {
     @ActivityScope
     @ContributesAndroidInjector(modules = [
         MainActivityBindingsModule::class,
+        MainActivityProvidesModule::class,
         FragmentBuilderModule::class,
         NavigationExecutorModule::class
     ])
